updated schema, set fixed length for ID variables, tested/working

+++ src/main/kotlin/net/maizegenetics/phgv2/utils/VariantLoadingUtils.kt
@@ -368,8 +368,8 @@ fun md5ToMurmur3(md5hash: String): Long {
 
 // This function skips the MD5 hash and directly creates an INT 64 hash
 // for use with hvcf and tiledb. Will we transition to this as hapID value?
-fun xxHash64(input: String): Long {
-    return LongHashFunction.xx().hashBytes(input.toByteArray(StandardCharsets.UTF_8))
+fun xxHash64(input: String): ULong {
+    return LongHashFunction.xx().hashBytes(input.toByteArray(StandardCharsets.UTF_8)).toULong()
 }
